+++ src/ScottPlot5/ScottPlot5 Demos/ScottPlot5 WinForms Demo/Demos/MultiAxis.cs
@@ -35,13 +35,13 @@ public partial class MultiAxis : Form, IDemoWindow
         var sig1 = formsPlot1.Plot.Add.Signal(data1);
         sig1.Axes.YAxis = YAxis1;
         YAxis1.Label.Text = "YAxis1";
-        YAxis1.Label.ForeColor = sig1.LineStyle.Color;
+        YAxis1.Label.ForeColor = sig1.Color;
 
         double[] data2 = ScottPlot.RandomDataGenerator.Generate.RandomWalk(count: 51, mult: 1000);
         var sig2 = formsPlot1.Plot.Add.Signal(data2);
         sig2.Axes.YAxis = YAxis2;
         YAxis2.Label.Text = "YAxis2";
-        YAxis2.Label.ForeColor = sig2.LineStyle.Color;
+        YAxis2.Label.ForeColor = sig2.Color;
 
         formsPlot1.Plot.AutoScale();
         formsPlot1.Plot.Zoom(.8, .7); // zoom out slightly